requirements.txt
anyio==3.6.2
APScheduler==3.9.1.post1
astor==0.8.1
attrs==22.1.0
bandit==1.7.4
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
darglint==1.8.1
docutils==0.19
eradicate==2.1.0
flake8==4.0.1
flake8-bandit==3.0.0
flake8-broken-line==0.5.0
flake8-bugbear==22.10.27
flake8-commas==2.1.0
flake8-comprehensions==3.10.1
flake8-debugger==4.1.2
flake8-docstrings==1.6.0
flake8-eradicate==1.4.0
flake8-isort==4.2.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.1
flake8-rst-docstrings==0.2.7
flake8-string-format==0.3.0
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
isort==5.10.1
kiwisolver==1.4.4
matplotlib==3.6.2
mccabe==0.6.1
mypy==0.990
mypy-extensions==0.4.3
numpy==1.23.4
packaging==21.3
pathspec==0.10.1
pbr==5.11.0
pep8-naming==0.13.2
Pillow==9.3.0
platformdirs==2.5.3
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.0a4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
restructuredtext-lint==1.4.0
rfc3986==1.5.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
stevedore==4.1.1
tomli==2.0.1
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
wemake-python-styleguide==0.17.0
